requirements.txt
azure-storage-blob
azure-identity